+++ proxy.ts
@@ -1,13 +1,10 @@
-import NextAuth from 'next-auth'
-import authConfig from '@/lib/auth.config'
+import { auth } from '@/lib/auth'
 import {
   DEFAULT_LOGIN_REDIRECT,
   apiAuthPrefix,
   authRoutes,
   publicRoutes,
 } from '@/lib/routes'
-
-const { auth } = NextAuth(authConfig)
 
 export default auth((req) => {
   const { nextUrl } = req
